+++ packages/cli/src/command.ts
@@ -1,11 +1,15 @@
 import { Argument, Command } from "commander";
 import { mkdirp } from "fs-extra";
 import { Config } from "@vastly/types";
+import chalk from "chalk";
 import { errorToString, isErrnoException } from "@vastly/utils";
 import type { PackageJson } from "type-fest";
+import latestVersion from "latest-version";
+
 import { readConfigFile, writeToConfigFile, getConfigFilePath } from "./util/config/files.js";
 import getGlobalPathConfig from "./util/config/files.js";
 import { defaultConfig } from "./util/config/defaults.js";
+import { type Stage } from "./types/index.js";
 
 const makeClient = await import("./util/client.js");
 const makeOutput = await import("./util/output/create-output.js");
@@ -18,6 +22,17 @@ const { CLIENT_API_URL } = process.env;
 export async function makeProgram(program: Command, pkg: PackageJson) {
   const options = program.opts();
   const output = makeOutput.default({ stream: process.stderr, debugEnabled: options.debug });
+
+  try {
+    const lv = await latestVersion("@vastly/cli");
+    if (pkg.version && lv !== pkg.version) {
+      output.log(
+        `${chalk.green("Update available!")} ${chalk.red(pkg.version)} -> ${chalk.yellowBright.bold(lv)} \nRun ${chalk.cyan.inverse("pnpm add -g @vastly/cli")} to update \n`,
+      );
+    }
+  } catch (e) {
+    return 1;
+  }
 
   // Make sure global config dir exists
   try {
@@ -59,8 +74,9 @@ export async function makeProgram(program: Command, pkg: PackageJson) {
     stderr: process.stderr,
     output,
     config,
-    apiUrl: CLIENT_API_URL || '',
+    apiUrl: CLIENT_API_URL || "",
   });
+  const allowableStages: Stage[] = ["sandbox", "local", "uat", "prod", "production"];
 
   program
     .name("vastly")
@@ -89,7 +105,17 @@ export async function makeProgram(program: Command, pkg: PackageJson) {
     .description("Manage your environment variables")
     .addArgument(new Argument("<action>", "Env options").choices(["add", "delete", "pull"]))
     .option("-a, --all", "Pull all environment variables")
-    .action(async (action, options) => {
+    .requiredOption("-st, --stage <stage>", "Pass a stage argument to add, delete, or pull from")
+    .action(async (action: string, options: { stage: Stage; all?: boolean }) => {
+      if (options.stage && !allowableStages.includes(options.stage)) {
+        console.log(
+          `${chalk.red(`Invalid stage! Your choices are ${chalk.inverse(allowableStages.join(", "))}`)}`,
+        );
+        process.exit(1);
+      }
+      if (options.stage === "production") {
+        options.stage = "prod";
+      }
       const func = (await import("./commands/env/index.js")).default;
       await func(client, action, options);
     });
@@ -98,7 +124,7 @@ export async function makeProgram(program: Command, pkg: PackageJson) {
     .command("whoami")
     .description("Display the username of the currently logged in user")
     .option("-t, --token", "Returns your current token from user config")
-    .action(async (_args, options) => {
+    .action(async (_args: any, options) => {
       const func = (await import("./commands/whoami.js")).default;
 
       await func(client, options.opts());