+++ packages/cli/src/commands/env/index.ts
@@ -4,21 +4,26 @@ import addEnv from "./add.js";
 import deleteEnv from "./delete.js";
 import pullEnv from "./pull.js";
 import pullAllEnv from "./pull-all.js";
+import { type Stage } from "../../types/index.js";
 
-export default async function determineEnvCommand(client: Client, action: string, options: any) {
+export default async function determineEnvCommand(
+  client: Client,
+  action: string,
+  options: { stage: Stage; all?: boolean },
+) {
   try {
     switch (action) {
       case "add":
-        await addEnv(client);
+        await addEnv(client, options.stage);
         break;
       case "delete":
-        await deleteEnv(client);
+        await deleteEnv(client, options.stage);
         break;
       case "pull":
         if (options.all) {
-          await pullAllEnv(client);
+          await pullAllEnv(client, options.stage);
         } else {
-          await pullEnv(client);
+          await pullEnv(client, options.stage);
         }
         break;
       default: